Moved integration test to examples
