+++ packages/crystallis_generator/example/copy_methods/pubspec.yaml
@@ -1,5 +1,5 @@
-name: integration
-description: These are integration tests for crystallis_generator.
+name: copy_methods_example
+description: These are copy methods' integration examples for crystallis_generator.
 version: 1.0.0
 
 publish_to: none
@@ -10,11 +10,12 @@ environment:
 resolution: workspace
 
 dependencies:
+  crystallis:
+    path: ../../
   path: ^1.9.0
 
 dev_dependencies:
   build_runner: ^2.12.2
-  crystallis: any
   crystallis_generator:
     path: ../../
   test: ^1.30.0
